refactor: move getProjectComponents to core for VS
